backend/requirements.txt
# Flask and Extensions
flask>=2.0.1
flask-sqlalchemy>=3.0.2
flask-migrate>=4.0.4
flask-cors>=4.0.0
flask-jwt-extended>=4.5.2
flask-login>=0.6.2
flask-mail>=0.9.1
flask-marshmallow>=0.15.0
flask-socketio>=5.3.6
marshmallow-sqlalchemy>=0.29.0

# Database
sqlalchemy>=2.0.0
alembic>=1.11.1
psycopg2-binary>=2.9.9
asyncpg>=0.29.0

# WebSocket
websockets>=11.0.3
python-socketio>=5.10.0
eventlet>=0.33.3

# Security
bcrypt>=4.0.1
pyjwt>=2.8.0
python-jose>=3.3.0
passlib>=1.7.4

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
pytest-flask>=1.2.0
pytest-mock>=3.10.0
nest-asyncio>=1.6.0
uvloop>=0.19.0; platform_system != "Windows"

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.8.5
redis>=4.6.0
email-validator>=2.0.0

# Development
black>=23.12.1
flake8>=7.0.0
mypy>=1.8.0
isort>=5.13.2

# Core dependencies
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
email-validator>=2.1.0
python-dotenv>=1.0.0

# Database
SQLAlchemy>=2.0.23
alembic>=1.12.1
psycopg2-binary>=2.9.9
asyncpg>=0.29.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.1
pytest-cov>=4.1.0

# Utilities
python-dateutil>=2.8.2
requests>=2.31.0
aiofiles>=23.2.1
Pillow>=10.1.0
python-magic>=0.4.27

# WebSocket
websockets>=12.0
redis>=5.0.1

# Celery
celery>=5.3.6
flower>=2.0.1

# Logging
structlog>=23.2.0
